move stop_time
